Add departments into migrations (#93)
The environment variables for the backend were previously defined in
the docker-compose.yml. However, this is a bad practice. This is because
the environment vars will override whatever is in the .env file. This
means that for testing the settings needed to be manually re-define essentially making the settings object useless.

.env.development and .env.test have been committed to the repo. 

The settings are filled as follows: if no .env* files exist, then use shell variables(they override
everything), if a .env file(not commited) exists pull the vars from
that, if ENVIRONMENT=test is passed in to a command then use the .env.test file,
otherwise use .env.development.

There was a variable called BACKEND_DOCKER_DEFAULT_PLATFORM. In order to use
this, manually set this value in your shell. It doesn't get loaded into
settings and thus shouldn't need to be in our .env file.

---------

Co-authored-by: sbailleul <44065690+sbailleul@users.noreply.github.com>